수정: 2차 모범답안 내용으로 변경함 1. 앞으로 Spring Boot으로 마이그레이션할때는 초반 모범답안과 비슷한게 추후에 수정이 용이하다고 함. 2. 모범답안과 내 코드의 차이는 확인하였음   - 모범 답안은 FileRepo에서 File을 찾은 뒤, 따로 Map에 저장하지 않음. File 자체로만 컨트롤 함.     - 나의 FileRepo는 File을 찾은 뒤, Map에 추가적으로 저장하여 조회에 사용함.   - Entity의 update메서드에서 update방식이 다름.     - 모범답안은 뭐라도 수정되는지 체크하고(bool 변수 사용), update날짜를 변경     - 나의 코드는 이 부분이 없음. 무조건 update날짜를 변경 중

+++ src/main/java/com/sprint/mission/discodeit/service/file/FileChannelService.java
@@ -1,21 +1,23 @@
-package com.sprint.mission.discodeit.repository.file;
+package com.sprint.mission.discodeit.service.file;
 
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.repository.ChannelRepository;
+import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.service.ChannelService;
 
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 
-public class FileChannelRepository implements ChannelRepository {
+public class FileChannelService implements ChannelService {
     private final Path DIRECTORY;
     private final String EXTENSION = ".ser";
 
-    public FileChannelRepository() {
+    public FileChannelService() {
         this.DIRECTORY = Paths.get(System.getProperty("user.dir"), "file-data-map", Channel.class.getSimpleName());
         if (Files.notExists(DIRECTORY)) {
             try {
@@ -31,7 +33,8 @@ public class FileChannelRepository implements ChannelRepository {
     }
 
     @Override
-    public Channel save(Channel channel) {
+    public Channel create(ChannelType type, String name, String description) {
+        Channel channel = new Channel(type, name, description);
         Path path = resolvePath(channel.getId());
         try (
                 FileOutputStream fos = new FileOutputStream(path.toFile());
@@ -41,13 +44,14 @@ public class FileChannelRepository implements ChannelRepository {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+
         return channel;
     }
 
     @Override
-    public Optional<Channel> findById(UUID id) {
+    public Channel find(UUID channelId) {
         Channel channelNullable = null;
-        Path path = resolvePath(id);
+        Path path = resolvePath(channelId);
         if (Files.exists(path)) {
             try (
                     FileInputStream fis = new FileInputStream(path.toFile());
@@ -58,7 +62,9 @@ public class FileChannelRepository implements ChannelRepository {
                 throw new RuntimeException(e);
             }
         }
-        return Optional.ofNullable(channelNullable);
+
+        return Optional.ofNullable(channelNullable)
+                .orElseThrow(() -> new NoSuchElementException("Channel with id " + channelId + " not found"));
     }
 
     @Override
@@ -83,14 +89,42 @@ public class FileChannelRepository implements ChannelRepository {
     }
 
     @Override
-    public boolean existsById(UUID id) {
-        Path path = resolvePath(id);
-        return Files.exists(path);
+    public Channel update(UUID channelId, String newName, String newDescription) {
+        Channel channelNullable = null;
+        Path path = resolvePath(channelId);
+        if (Files.exists(path)) {
+            try (
+                    FileInputStream fis = new FileInputStream(path.toFile());
+                    ObjectInputStream ois = new ObjectInputStream(fis)
+            ) {
+                channelNullable = (Channel) ois.readObject();
+            } catch (IOException | ClassNotFoundException e) {
+                throw new RuntimeException(e);
+            }
+        }
+
+        Channel channel = Optional.ofNullable(channelNullable)
+                .orElseThrow(() -> new NoSuchElementException("Channel with id " + channelId + " not found"));
+        channel.update(newName, newDescription);
+
+        try(
+                FileOutputStream fos = new FileOutputStream(path.toFile());
+                ObjectOutputStream oos = new ObjectOutputStream(fos)
+        ) {
+            oos.writeObject(channel);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        return channel;
     }
 
     @Override
-    public void deleteById(UUID id) {
-        Path path = resolvePath(id);
+    public void delete(UUID channelId) {
+        Path path = resolvePath(channelId);
+        if (Files.notExists(path)) {
+            throw new NoSuchElementException("Channel with id " + channelId + " not found");
+        }
         try {
             Files.delete(path);
         } catch (IOException e) {
